Motor, ImageProcess

+++ CODE/Inc/User_OSTU.h
@@ -13,7 +13,7 @@
 #define IMAGEH  MT9V03X_H
 #define IMAGEW  MT9V03X_W
 
-uint8_t GetOSTU(uint8_t tmImage[IMAGEH][IMAGEW]);
+uint16_t GetOSTU(uint8_t tmImage[IMAGEH][IMAGEW]);
 void BinaryImage(uint8_t tmImage[IMAGEH][IMAGEW],uint8_t ThresholdV);
 //void GyroReadByte(void);
